add new case for api/txs/hash

+++ src/main/java/microservices/txn/models/TransactionInfo.java
@@ -15,6 +15,6 @@ public class TransactionInfo {
     public int epochNo;
     public String status;
     public int confirmation;
-    public int fee;
-    public int totalOutput;
+    public long fee;
+    public long totalOutput;
 }
